Change quantity variable type to double

+++ app/src/main/java/com/example/android/bakingapp/models/Ingredient.java
@@ -11,7 +11,7 @@ public class Ingredient implements Parcelable
 
     @SerializedName("quantity")
     @Expose
-    private Integer quantity;
+    private double quantity;
     @SerializedName("measure")
     @Expose
     private String measure;
@@ -36,7 +36,7 @@ public class Ingredient implements Parcelable
             ;
 
     protected Ingredient(Parcel in) {
-        this.quantity = ((Integer) in.readValue((Integer.class.getClassLoader())));
+        this.quantity = ((Double) in.readValue((Double.class.getClassLoader())));
         this.measure = ((String) in.readValue((String.class.getClassLoader())));
         this.ingredient = ((String) in.readValue((String.class.getClassLoader())));
     }
@@ -44,7 +44,7 @@ public class Ingredient implements Parcelable
     public Ingredient() {
     }
 
-    public Integer getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
